refactor: global.config package
